Moved examples to another module
