+++ bigtable/go.mod
@@ -5,15 +5,20 @@ go 1.13
 require (
 	cloud.google.com/go v0.83.0 // indirect
 	cloud.google.com/go/bigtable v1.10.1
+	github.com/fsnotify/fsnotify v1.4.9 // indirect
 	github.com/golang/protobuf v1.5.2
 	github.com/google/btree v1.0.1
 	github.com/google/go-cmp v0.5.6
+	github.com/kr/pretty v0.2.0 // indirect
+	github.com/onsi/ginkgo v1.11.0 // indirect
+	github.com/onsi/gomega v1.7.1 // indirect
 	github.com/syndtr/goleveldb v1.0.0
-	github.com/testcontainers/testcontainers-go v0.11.1 // indirect
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 	golang.org/x/sys v0.0.0-20210603125802-9665404d3644 // indirect
 	google.golang.org/api v0.47.0
 	google.golang.org/genproto v0.0.0-20210602131652-f16073e35f0c
 	google.golang.org/grpc v1.38.0
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 	rsc.io/binaryregexp v0.2.0
 )
