--- docker.sh
@@ -1,3 +1,0 @@
-# TODO: what are the actual image tags (including repo)?
-docker build --target=cbtemulator -t cbtemulator .
-docker build --target=gcsemulator -t gcsemulator .
