+++ examples/bigtable/bigtable.go
@@ -1,19 +1,4 @@
-/*
-Copyright 2016 Google LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package bttest_test
+package bigtable
 
 import (
 	"cloud.google.com/go/bigtable"
@@ -27,7 +12,7 @@ import (
 	"testing"
 )
 
-func TestExampleLocalServer(t *testing.T) {
+func LocalServer(t *testing.T) {
 	srv, err := bttest.NewServer("localhost:0")
 	if err != nil {
 		t.Fatal(err)
@@ -39,10 +24,10 @@ func TestExampleLocalServer(t *testing.T) {
 	}
 }
 
-func TestExampleContainerServer(t *testing.T) {
+func ContainerServer(t *testing.T) {
 	ctx := context.Background()
 	req := testcontainers.ContainerRequest{
-		Image:        "fullstorydev/cbtemulator:latest",
+		Image:        "fullstorydev/cbtemulator:b5414bf-dirty",
 		ExposedPorts: []string{"9000"},
 	}
 	c, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
