+++ examples/go.mod
@@ -1,0 +1,8 @@
+module github.com/fullstorydev/emulators/storage
+
+go 1.13
+
+require (
+	github.com/fullstorydev/emulators/bigtable v0.0.0-20210813131059-a07c93f50fe7 // indirect
+	github.com/testcontainers/testcontainers-go v0.11.1 // indirect
+)
