+++ storage/go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/bluele/gcache v0.0.2
 	github.com/golang/protobuf v1.5.2
 	github.com/google/btree v1.0.1
-	github.com/testcontainers/testcontainers-go v0.11.1
+	github.com/pkg/errors v0.9.1 // indirect
 	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c
 	golang.org/x/sys v0.0.0-20210603125802-9665404d3644 // indirect
 	google.golang.org/api v0.47.0
